refactor(chat): move Stop into the assistant bubble
The run belongs to that bubble, so the control that ends it does too.
The header keeps the Working badge; ChatTranscript only relays state
and the stop request between the run and the message that owns it.
