+++ gui/views/chat/ChatWorkflow.vue
@@ -46,17 +46,6 @@
               messages.filter(m => m.role === 'assistant').length === 1 ? '' : 's'
             }}
           </span>
-          <button
-            v-if="run.isRunning.value"
-            class="stop-button"
-            :disabled="run.isStopping.value"
-            @click="run.requestStop()"
-            type="button"
-            :title="stopTitle"
-          >
-            <font-awesome-icon :icon="stopIcon" />
-            <span>{{ run.isStopping.value ? 'Stopping…' : 'Stop' }}</span>
-          </button>
         </div>
         <div class="header-right">
           <button
@@ -92,6 +81,8 @@
         :workflow-name="workflow?.display_name"
         :split-sentences="splitSentences"
         :is-injected-sentence="isInjectedSentence"
+        :stopping="run.isStopping.value"
+        @stop="run.requestStop()"
       />
 
       <ChatComposer
@@ -119,7 +110,7 @@
 <script setup>
 import { ref, computed, inject, watch, onMounted, onBeforeUnmount, nextTick } from 'vue'
 import { FontAwesomeIcon } from '@fortawesome/vue-fontawesome'
-import { faAngleLeft, faPlus, faStop } from '@fortawesome/free-solid-svg-icons'
+import { faAngleLeft, faPlus } from '@fortawesome/free-solid-svg-icons'
 import ChatSidebar from './ChatSidebar.vue'
 import ChatTranscript from './ChatTranscript.vue'
 import ChatComposer from './ChatComposer.vue'
@@ -138,7 +129,6 @@ const globalConfig = inject('mcpGlobalConfig')
 const availableServers = inject('mcpAvailableServers', ref([]))
 const backIcon = faAngleLeft
 const plusIcon = faPlus
-const stopIcon = faStop
 
 // --- Persistent per-workflow chat state -------------------------------------
 // useChatSession owns messages / sessionId / historyEnabled / selectedRag
@@ -274,12 +264,6 @@ function _resumeRunInFlight() {
 onBeforeUnmount(run.stop)
 
 // --- Workflow-derived UI bits ----------------------------------------------
-const stopTitle = computed(() =>
-  run.isStopping.value
-    ? 'Stopping. The agent loop takes a few seconds to unwind.'
-    : 'Stop this run. Anything already created in CALDERA stays.'
-)
-
 const examplePrompts = computed(() => props.workflow?.example_prompts || [])
 const composerPlaceholder = computed(() => {
   // The box is closed, but the page is not: say so.
@@ -572,30 +556,6 @@ function clearTranscript() {
   color: #c9b8ff;
   background-color: rgba(139, 92, 246, 0.12);
   border-color: rgba(169, 112, 255, 0.28);
-}
-.stop-button {
-  display: inline-flex;
-  align-items: center;
-  gap: 0.4rem;
-  height: 28px;
-  padding: 0 0.7rem;
-  border-radius: 6px;
-  font-size: 0.82rem;
-  line-height: 1;
-  cursor: pointer;
-  color: #ffd7c2;
-  background-color: rgba(239, 108, 68, 0.14);
-  border: 1px solid rgba(239, 108, 68, 0.5);
-  transition: color 0.15s ease, background-color 0.15s ease, border-color 0.15s ease;
-}
-.stop-button:hover:not(:disabled) {
-  color: #ffffff;
-  background-color: rgba(239, 108, 68, 0.28);
-  border-color: #ef6c44;
-}
-.stop-button:disabled {
-  opacity: 0.5;
-  cursor: not-allowed;
 }
 .status-dot {
   width: 7px;
